Backport HIVE-4973. Disable add jar, dfs and add file based on config in hive-site.xml

+++ service/src/java/org/apache/hive/service/cli/operation/ExecuteStatementOperation.java
@@ -22,6 +22,9 @@ package org.apache.hive.service.cli.operation;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.hadoop.hive.conf.HiveConf;
+import org.apache.hadoop.hive.conf.HiveConf.ConfVars;
+import org.apache.hive.service.cli.HiveSQLException;
 import org.apache.hive.service.cli.OperationType;
 import org.apache.hive.service.cli.session.HiveSession;
 
@@ -40,20 +43,38 @@ public abstract class ExecuteStatementOperation extends Operation {
   }
 
   public static ExecuteStatementOperation newExecuteStatementOperation(
-      HiveSession parentSession, String statement, Map<String, String> confOverlay) {
+      HiveSession parentSession, String statement, Map<String, String> confOverlay) throws HiveSQLException {
     String[] tokens = statement.trim().split("\\s+");
     String command = tokens[0].toLowerCase();
 
+    ExecuteStatementOperation newExecOP;
+    HiveConf hiveConf = parentSession.getHiveConf();
+    boolean allowExternalExec = hiveConf.getBoolVar(HiveConf.ConfVars.HIVE_SERVER2_AUTHZ_EXTERNAL_EXEC);
     if ("set".equals(command)) {
-      return new SetOperation(parentSession, statement, confOverlay);
+       newExecOP = new SetOperation(parentSession, statement, confOverlay);
     } else if ("dfs".equals(command)) {
-      return new DfsOperation(parentSession, statement, confOverlay);
+      if (allowExternalExec) {
+        newExecOP = new DfsOperation(parentSession, statement, confOverlay);
+      } else {
+        throw new HiveSQLException("Insufficient privileges to execute dfs", "42000");
+      }
     } else if ("add".equals(command)) {
-      return new AddResourceOperation(parentSession, statement, confOverlay);
+      if (allowExternalExec) {
+        newExecOP = new AddResourceOperation(parentSession, statement, confOverlay);
+      } else {
+        throw new HiveSQLException("Insufficient privileges to execute add", "42000");
+      }
     } else if ("delete".equals(command)) {
-      return new DeleteResourceOperation(parentSession, statement, confOverlay);
+      if (allowExternalExec) {
+        newExecOP = new DeleteResourceOperation(parentSession, statement, confOverlay);
+      } else {
+        throw new HiveSQLException("Insufficient privileges to execute delete", "42000");
+      }
+
     } else {
-      return new SQLOperation(parentSession, statement, confOverlay);
+      newExecOP = new SQLOperation(parentSession, statement, confOverlay);
     }
-  }
+
+    return newExecOP;
+ }
 }
